Improve organisation of tests
